+++ CarboPlanner/src/components/ManualAddFoodModal.tsx
